+++ front/components/spaces/SpaceCategoriesList.tsx
@@ -26,8 +26,8 @@ import type { CellContext } from "@tanstack/react-table";
 import type { ComponentType } from "react";
 import React from "react";
 
+import { ACTION_BUTTONS_CONTAINER_ID } from "@app/components/spaces/SpacePageTools";
 import { SpaceSearchContext } from "@app/components/spaces/SpaceSearchContext";
-import { ACTION_BUTTONS_CONTAINER_ID } from "@app/components/spaces/SpaceSearchLayout";
 import { useActionButtonsPortal } from "@app/hooks/useActionButtonsPortal";
 import { CATEGORY_DETAILS } from "@app/lib/spaces";
 import { useSpaceInfo } from "@app/lib/swr/spaces";
