+++ front/components/spaces/SpaceDataSourceViewContentList.tsx
@@ -55,8 +55,8 @@ import {
 } from "@app/components/spaces/ContentActions";
 import { EditSpaceManagedDataSourcesViews } from "@app/components/spaces/EditSpaceManagedDatasourcesViews";
 import { FoldersHeaderMenu } from "@app/components/spaces/FoldersHeaderMenu";
+import { ACTION_BUTTONS_CONTAINER_ID } from "@app/components/spaces/SpacePageTools";
 import { SpaceSearchContext } from "@app/components/spaces/SpaceSearchContext";
-import { ACTION_BUTTONS_CONTAINER_ID } from "@app/components/spaces/SpaceSearchLayout";
 import { WebsitesHeaderMenu } from "@app/components/spaces/WebsitesHeaderMenu";
 import { useActionButtonsPortal } from "@app/hooks/useActionButtonsPortal";
 import { getVisualForDataSourceViewContentNode } from "@app/lib/content_nodes";
