+++ front/components/spaces/SpaceResourcesList.tsx
@@ -43,9 +43,12 @@ import type { DataSourceIntegration } from "@app/components/spaces/AddConnection
 import { AddConnectionMenu } from "@app/components/spaces/AddConnectionMenu";
 import { EditSpaceManagedDataSourcesViews } from "@app/components/spaces/EditSpaceManagedDatasourcesViews";
 import { EditSpaceStaticDatasourcesViews } from "@app/components/spaces/EditSpaceStaticDatasourcesViews";
+import { ACTION_BUTTONS_CONTAINER_ID } from "@app/components/spaces/SpacePageTools";
+import { SpaceSearchContext } from "@app/components/spaces/SpaceSearchContext";
 import { UsedByButton } from "@app/components/spaces/UsedByButton";
 import { useTheme } from "@app/components/sparkle/ThemeContext";
 import { ViewFolderAPIModal } from "@app/components/ViewFolderAPIModal";
+import { useActionButtonsPortal } from "@app/hooks/useActionButtonsPortal";
 import {
   getConnectorProviderLogoWithFallback,
   isConnectorPermissionsEditable,
@@ -57,9 +60,6 @@ import {
   useSpaceDataSourceViewsWithDetails,
 } from "@app/lib/swr/spaces";
 import { classNames } from "@app/lib/utils";
-import { SpaceSearchContext } from "@app/components/spaces/SpaceSearchContext";
-import { ACTION_BUTTONS_CONTAINER_ID } from "@app/components/spaces/SpaceSearchLayout";
-import { useActionButtonsPortal } from "@app/hooks/useActionButtonsPortal";
 
 export interface RowData {
   dataSourceView: DataSourceViewsWithDetails;
