+++ front/components/spaces/search/columns.tsx
@@ -11,49 +11,57 @@ type RowData = DataSourceViewContentNode & {
   menuItems?: MenuItem[];
 };
 
-export function makeColumnsForSearchResults(): ColumnDef<RowData>[] {
-  const columns: ColumnDef<RowData, any>[] = [];
-  columns.push({
-    header: "Name",
-    accessorKey: "title",
-    id: "title",
-    sortingFn: "text", // built-in sorting function case-insensitive
-    cell: (info: CellContext<RowData, string>) => (
-      <DataTable.CellContent icon={info.row.original.icon}>
-        <span>{info.getValue()}</span>
-      </DataTable.CellContent>
-    ),
-  });
-
-  columns.push({
-    header: "Last updated",
-    id: "lastUpdatedAt",
-    accessorKey: "lastUpdatedAt",
-    meta: {
-      className: "w-20",
-    },
-    cell: (info: CellContext<RowData, number>) => (
-      <DataTable.BasicCellContent
-        className="justify-end"
-        label={
-          info.getValue()
-            ? formatTimestampToFriendlyDate(info.getValue(), "compact")
-            : "-"
-        }
-      />
-    ),
-  });
-
-  columns.push({
-    id: "actions",
-    meta: {
-      className: "flex justify-end items-center",
-    },
-    cell: (info) =>
-      info.row.original.menuItems && (
-        <DataTable.MoreButton menuItems={info.row.original.menuItems} />
+export function makeColumnsForSearchResults(): ColumnDef<RowData, any>[] {
+  return [
+    {
+      header: "Name",
+      accessorKey: "title",
+      id: "title",
+      enableSorting: false,
+      cell: (info: CellContext<RowData, string>) => (
+        <DataTable.CellContent icon={info.row.original.icon}>
+          <span>{info.getValue()}</span>
+        </DataTable.CellContent>
       ),
-  });
-
-  return columns;
+    },
+    {
+      header: "Location",
+      accessorKey: "location",
+      id: "location",
+      enableSorting: false,
+      cell: (info: CellContext<RowData, string>) => (
+        <DataTable.BasicCellContent label={info.getValue()} />
+      ),
+    },
+    {
+      header: "Last updated",
+      id: "lastUpdatedAt",
+      accessorKey: "lastUpdatedAt",
+      enableSorting: false,
+      meta: {
+        className: "w-20",
+      },
+      cell: (info: CellContext<RowData, number>) => (
+        <DataTable.BasicCellContent
+          className="justify-end"
+          label={
+            info.getValue()
+              ? formatTimestampToFriendlyDate(info.getValue(), "compact")
+              : "-"
+          }
+        />
+      ),
+    },
+    {
+      id: "actions",
+      enableSorting: false,
+      meta: {
+        className: "flex justify-end items-center",
+      },
+      cell: (info) =>
+        info.row.original.menuItems && (
+          <DataTable.MoreButton menuItems={info.row.original.menuItems} />
+        ),
+    },
+  ];
 }
